Refactor code to pull PQ fetching logic into single class
